python_nodes/pub_test_python/requirements.txt
eclipse-zenoh==1.0.4
numpy

